Added outlier removal and data smoothing
Added outlier removal and data smoothing. any value that is more than 3
standard deviations away from the mean of the last 10 readings is
replaced by the average of the last 10 readings. Data is now smoothed
using a moving average (aka rolling average) of the last 10 readings.
Also added a test program to demonstrate the smoothing and a utility to
dump the binary array stored in a file of the last 10 readings.